# backend/src/main/java/com/chamuditha/tournament/auth/AuthenticationResponse.java
package com.chamuditha.tournament.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;
}